+++ axum-core/src/extract/default_body_limit.rs
@@ -72,7 +72,7 @@ use tower_layer::Layer;
 /// [`RequestBodyLimit`]: tower_http::limit::RequestBodyLimit
 /// [`RequestExt::with_limited_body`]: crate::RequestExt::with_limited_body
 /// [`RequestExt::into_limited_body`]: crate::RequestExt::into_limited_body
-#[derive(Debug, Clone)]
+#[derive(Debug, Clone, Copy)]
 #[must_use]
 pub struct DefaultBodyLimit {
     kind: DefaultBodyLimitKind,
@@ -116,7 +116,7 @@ impl DefaultBodyLimit {
     /// [`Bytes`]: bytes::Bytes
     /// [`Json`]: https://docs.rs/axum/0.7/axum/struct.Json.html
     /// [`Form`]: https://docs.rs/axum/0.7/axum/struct.Form.html
-    pub fn disable() -> Self {
+    pub const fn disable() -> Self {
         Self {
             kind: DefaultBodyLimitKind::Disable,
         }
@@ -149,7 +149,7 @@ impl DefaultBodyLimit {
     /// [`Bytes::from_request`]: bytes::Bytes
     /// [`Json`]: https://docs.rs/axum/0.7/axum/struct.Json.html
     /// [`Form`]: https://docs.rs/axum/0.7/axum/struct.Form.html
-    pub fn max(limit: usize) -> Self {
+    pub const fn max(limit: usize) -> Self {
         Self {
             kind: DefaultBodyLimitKind::Limit(limit),
         }